refactor(scripts): extract the Bindings type parse into scripts/bindings.ts
`Bindings` is a type with no runtime form, so everything that needs the
binding list parses src/index.ts textually. That parse had been written
twice — once in build-manifest.ts and once again in
tests/config-registry.test.ts, whose copy was commented "same textual
parse build-manifest.ts performs". Two parsers that must agree is the
drift class the config registry exists to kill.

Behavior-identical move: parseBindings() is now pure over source text
(so a generator can be tested against a fixture) with a
readBindingsSource() helper for the real file. release-manifest.json is
byte-identical.
